@@ -2,25 +2,26 @@
 
 """The setup script."""
 
+import pathlib
+
 from setuptools import find_packages, setup
 
 with open('requirements.txt') as f:
     requirements = f.read().strip().split('\n')
 
-with open('README.md') as f:
-    long_description = f.read()
+long_description = pathlib.Path('README.md').read_text()
 setup(
     maintainer='Xdev',
     maintainer_email='xdev@ucar.edu',
-    python_requires='>=3.7',
+    python_requires='>=3.10',
     classifiers=[
         'Development Status :: 2 - Pre-Alpha',
         'License :: OSI Approved :: Apache Software License',
         'Natural Language :: English',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.7',
-        'Programming Language :: Python :: 3.8',
-        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
+        'Programming Language :: Python :: 3.12',
         'Topic :: Scientific/Engineering',
         'Operating System :: OS Independent',
         'Intended Audience :: Science/Research',
