@@ -1,4 +1,3 @@
-import typing
 from pathlib import Path
 
 import typer
@@ -23,7 +22,7 @@ def start(
     port: int = typer.Option(
         8888,
         help=(
-            '''The local port the remote notebook server will be forwarded to. If not specified, defaults to 8888.'''
+            """The local port the remote notebook server will be forwarded to. If not specified, defaults to 8888."""
         ),
         show_default=True,
     ),
@@ -40,7 +39,7 @@ def start(
     notebook: str = typer.Option(
         None,
         show_default=True,
-        help='''The absolute path of the notebook to load on the remote host. `--notebook-dir` and `--notebook` are mutually exclusive.''',
+        help="""The absolute path of the notebook to load on the remote host. `--notebook-dir` and `--notebook` are mutually exclusive.""",
     ),
     port_forwarding: bool = typer.Option(
         True, show_default=True, help='Whether to set up SSH port forwarding or not.'
@@ -53,7 +52,7 @@ def start(
         file_okay=True,
         show_default=True,
         help=(
-            '''Selects a file from which the identity (private key) for public key authentication is read.'''
+            """Selects a file from which the identity (private key) for public key authentication is read."""
         ),
     ),
     launch_command: str = typer.Option(
@@ -72,7 +71,7 @@ def start(
         show_default=True,
         help='Which remote shell binary to use.',
     ),
-    version: typing.Optional[bool] = typer.Option(
+    version: bool | None = typer.Option(
         None,
         '--version',
         callback=version_callback,
